+++ server/src/main/java/io/littlehorse/server/monitoring/health/ServerHealthState.java
@@ -1,7 +1,6 @@
 package io.littlehorse.server.monitoring.health;
 
 import io.littlehorse.common.LHServerConfig;
-import io.littlehorse.server.monitoring.InProgressRestoration;
 import io.littlehorse.server.monitoring.StandbyStoresOnInstance;
 import java.util.ArrayList;
 import java.util.List;
